backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
pydub
SpeechRecognition
python-multipart
psycopg2-binary
google-genai
google-api-core
fpdf2
pandas
openpyxl
numpy
librosa
scikit-learn
PyMuPDF
python-docx
qdrant-client
langchain-google-genai
langchain-core
langchain-text-splitters
python-magic
supabase
python-jose[cryptography]
